Bresenham line

+++ point.go
@@ -2,6 +2,6 @@ package rasta
 
 // Point - Rasta Point
 type Point struct {
-	X float64 `json:"x"`
-	Y float64 `json:"y"`
+	X int `json:"x"`
+	Y int `json:"y"`
 }
